feat(serial): only try connect to available ports
For a given serial communications port, there are two main reasons why a
DropBot connection attempt may fail: 1) the port/device is busy, e.g.,
port is already opened in another process; or 2) the serial port does not
have a DropBot connected to it.

Prior to this commit, all serial ports were considered as candidates for
connection attempts.  However, this unnecessarily attempted to connect to
ports that were not available, corresponding to case (1) above.

As of this commit, only ports that are available for connection will be
considered.