chore(conda): require micro version in git tag
Require micro version to be explicitly set as part of the `git` tag.

Prior to this commit, the micro version was inferred as the number of
commits since the most recent annotated `git` tag.

To better align with explicit semantic versioning, this commit requires
the micro version to be explicitly set as part of the `git` tag.  Builds
on commits that are not annotated tags result in `.post<describe number>`
appended to the `git` tag version, where `<describe number>` corresponds
to the number of commits since the most recent annotated `git` tag.