fix(#23): keep HV on after chip inserted
High voltage is disabled (i.e., the low short-detection voltage is
selected) if the chip is not inserted regardless of the *high voltage
output enable* setting (i.e., `HV_OUTPUT_SELECT_PIN`) due to a hardware
interlock (i.e., the active low output enable `OE_PIN`).

If a chip is inserted when the high voltage output enable is set, the high
voltage should be turned on accordingly.  However, prior to this commit,
this does not happen.

As a workaround, this commit forces the high voltage to be turned off and
on, which seems to ensure high voltage is actually turned on.

TODO
----

Find out the root cause and find a more permanent fix.

See #23 for more information.