fix(#23): fix selection of HV output
If high voltage output is selected (as opposed to low short detection
voltage), and high voltage output is enabled (i.e., the hardware interlock
`OE_PIN` is pulled LOW and the boost converter is on), the high voltage
should be connected to the switching boards.  However, prior to this
commit, this does not happen.

As a workaround, this commit forces the high voltage to be turned off and
on, to ensure high voltage is actually connected to the switching boards.

See issue #23 for more information.