fix(hardware-tests): change order of test-channels replicates
Some channels fail (i.e., don't turn on) non-deterministically; therefore, we need to run replicates
to identify problematic switches. It appears that sweeping across all channels (i.e., from 0-119)
and then repeating is more sensitive to detecting bad switches than the alternative method of
repeating a single channel for a number of replicates and then moving on to the next. This
commit modifies the test to use the more sensitive method.