fix(feedback): fix capacitance method equation
Prior to this commit, the capacitance method equation did not include the
**actuation voltage** as a term in the denominator.  The equation was
derived from the [`dropbot.ProxyMixin.measure_capacitance`][1], but when
ported the **actuation voltage** term was inadvertently left out.
Empirically, the computed capacitance values were found to be off by
approximately a factor of 100, so the undocumented 0.15e-6 term was
modified to 0.15e-8, which seemed to solve the issue.  However, it turns
out that the "factor of ~100" was actually the missing actuation voltage
term and the 0.15e-6 term corresponded to the 0.15 uF capacitor term for
C16.

This commit fixes the `Node::capacitance` method equation to divide by
actuation voltage and documents the capacitance equation.

[1]: https://gitlab.com/sci-bots/dropbot.py/blob/129621a323153d3bccb97307de4e72cba85c0fb8/dropbot/proxy.py#L203