feat(firmware): add pack_drops function
Add `dropbot::drops::pack_drops` function to pack drops lists into single
array.

Each drop is encoded with a length ``n + 1``, where ``n`` is the number of
channels in the drop.  The first value in the encoded drop is ``n``,
followed by the drop channel numbers.

For example:

    [[0, 3, 7], [13, 2]]

is encoded as:

    [3, 0, 3, 7, 2, 13, 2]