fix(i2c): increase buffer size w/custom Wire lib
Prior to this commit, the firmware used the Teensyduino stock `Wire`
library, which does not provide any way to modify the I2C buffer size.
This results in a fixed I2C buffer size of 32 bytes.

In this commit, add a customized version of the stock `Wire`, replacing
all occurrences of `BUFFER_LENGTH` with `TWI_BUFFER_LENGTH`, and allowing
`TWI_BUFFER_LENGTH` to be overridden by a compile-time definition.

Fixes #31.