fix(events): fix `capacitance-exceeded` event
Since version 1.51, `capacitance-exceeded` events resulted in errors
similar to the following:

```python
DEBUG:base_node_rpc.queue:Stream packet contents do not describe an event: {"event": "capacitance-exceeded", "new_value": 3.38461e+125, "target": 1.13457e-11, "start": 3758096384, "end": 1034288511, "n_samples": 336656884
```

where the JSON message contained invalid values and the JSON message was
truncated, due to a misplaced "," typo in the arguments of a call to
`sprintf` (bug was introduced in version 1.51).

Also, prior to this commit, a `capacitance-exceeded` resulted in the
capacitance update _millisecond_ counter to be set _incorrectly_ to the
current _microsecond_ count.

In this commit, fix the `sprintf` call and set the capacitance update
counter using the _millisecond_ counter.