feat(threshold): add target_count to State
Prior to this commit, the a `capacitance-exceeded` event was triggered the
_first_ time the measured capacitance exceeded the target capacitance.
However, this can sometimes lead to an event being triggered too early due
to, e.g., a noisy capacitance reading that caused a temporary spike in
capacitance.

In this commit, require the measured capacitance to exceed target
capacitance for multiple consecutive readings before triggering a
`capacitance-exceeded` event.  The number of required consecutive readings
is set by the `target_count` attribute of the `dropbot::State` protocol
buffer.