fix(events): fix end time counters
Prior to this commit, the *end time* microseconds counter for capacitance
events was stored in the `now` variable, overwriting the millisecond
counter marking the start of the `loop()` method call, to avoid requiring
a new `uint32_t` instance.  However, this caused problems by downstream
code that relied on the `now` as the `loop()` start time.

In this commit, explicitly store the end time for capacitance events in a
new `uint32_t` variable.