fix(output-current): measure output RMS current
Prior to this commit, the output current was measured periodically and the
maximum value was compared to the output current threshold.  However,
without any further evidence, we suspect that RMS current is more
representative of what would cause the circuit to fail.

In this commit, add the `analog::measure_output_current_rms()` function to
measure the RMS output current and, when checking the output current
against the threshold value, compare the RMS output current over 20
samples instead of the maximum of 20 consecutive current readings.