refactor(output-enable): add chip_status_changed_
Prior to this commit, the `OutputEnableDebounce` managed callbacks itself.
This lead to code duplication for more general functionality already
provided by the `Signal` library.

In this commit, add `Node::chip_status_changed_` signal and refactor the
`Node` class and `OutputEnableDebounce` class to use the `Signal` instance
for managing callbacks based on the chip-inserted status.