fix(shorts): test all channels, even disabled
Prior to this commit, the short test only tested channels that were not
marked as disabled in the disabled channel mask.  However, the shorts test
_also **enabled**_ any channel where the measured voltage was above half
the reference voltage.  In the case of a disabled channel, no channel
would be turned on, potentially causing a false negative for a short on
the channel.

In this commit, force actuation of **all** channels during the shorts
test, _even_ channels marked as disabled in the disabled channels mask.
In the case where a channel was disabled, but passes the short test,
re-enable the channel.