feat(threads): add transaction_lock to proxy
Prior to this commit, multiple threads may have concurrently accessed a
proxy instance, potentially causing unexpected behaviour.

In this commit, add a `transaction_lock` attribute to the proxy.  This
lock may be used, e.g., for a thread to take control of the DropBot to
execute a sequence of steps _without being interrupted by another thread_.
Note that the lock is [_re-entrant_][1], i.e., functions that also require
the lock may be called while the lock is held.

[1]: https://stackoverflow.com/a/22885810/345236