feat(voltage): ensure high-voltage output on
Prior to this commit, the target high voltage setting could be set without
enabling the high voltage output.  This was not particularly intuitive.

In this commit, modify the `voltage` property setter to ensure
high-voltage output is turned on and enabled before setting target
voltage.

**Note:** the voltage can explicitly be set _without_ checking if the
high-voltage output is turned on (if needed) by using the `update_state()`
method.