feat(monitor): add async monitor coroutine
Add asynchronous DropBot monitor coroutine to establish and maintain a
DropBot connection.

If no DropBot is available or if the connection is lost, wait until a
DropBot is detected on one of the available serial ports and (re)connect.

DropBot signals are forwarded to the supplied `signals` namespace,
avoiding the need to manually connect signals after DropBot is
(re)connected.

DropBot connection is automatically closed when coroutine exits, e.g.,
when cancelled.