feat(py-drops): add Disconnect button to notebook
Prior to this commit, the `feat(py-drops)` notebook all cells could not be
run all at once.  First, cells requiring a DropBot connection did not wait
for the DropBot monitor to establish a connection.  Second, the last cell
(intended to demonstrate _how_ to disconnect from the DropBot) terminated
the DropBot connection.

In this commit:

 - Explicitly wait for a DropBot connection to be established before
   continuing on to cells requiring a connection.
 - Replace the last cell in the notebook to display a `Disconnect` button
   instead.  This allows all cells in the notebook to run without
   terminating the DropBot connection.