feat(analog): move Node::adc_ to analog::adc_
Prior to this commit, the `ADC` instance was stored as an attribute of the
`Node` object.

In this commit, deprecate `Node::adc_` attribute in favour of namespace
attribute, `analog::adc_`.  This makes it possible to implement the
following functions:

 - `save_config()`: Save current ADC configuration register values.
 - `load_config()`: Load saved ADC configuration register values.
 - `adc_context()`: Wrapper function to save/restore ADC configuration
   register contents.

Also, as part of this commit, wrap each analog measurement with
`adc_context()` to ensure each analog reading is executed with any ADC
configuration necessary (e.g., averaging, reference voltage, conversion
speed, etc.).