feat(channels): measure capacitance w/differential
Prior to this commit, the voltage used to calculate the chip load
capacitance was measured mode using single-ended mode (i.e., the feedback
pin referenced to ground).

In this commit, in an attempt to reduce measurement noise, measure the
voltage difference between the feedback pin and the 0.6 V virtual ground
to calculate the chip load capacitance instead.  The assumption is that
some of noise on the feedback pin will also be present on the virtual
ground pin, so measuring the difference should cancel such noise.