fix(event): fix current exceeded printf format
Prior to this commit, an incorrect format specifier for the `wall_time`
value in the  `high_side_current_exceeded_` signal callback lead to
undefined value output the all following keys, including `chip_load` and
`margin`.  This resulted in garbage values being reported in the event.

In this commit, fix the format specifier for the `wall_time` value in the
`high_side_current_exceeded_` signal callback.