fix(channels/capacitance): fix signed analog range
Prior to this commit, an incorrect range was used to normalize the ADC
signal amplitude during capacitance measurements.  The incorrect range was
_half_ the actual ADC range; since the maximum ADC value was used to
determine size of the range, but the range is twice the maximum value for
_signed_ readings.

In this commit, fix the range used to normalize the ADC signal amplitude
during capacitance measurements.