perf(proxy): cache number of channels on init
Prior to this commit, every access to the `number_of_channels` attribute
of the `ProxyMixin` class required a round-trip serial command
transaction (~8.5 ms on Windows 10).  However, the number of channels is
returned by the `ProxyMixin.initialize_switching_boards()` method

In this commit, cache the number of channels returned by the
`ProxyMixin.initialize_switching_boards()` method and return the cached
value as the `number_of_channels` attribute of the `ProxyMixin` class
Note that in the case where the number of channels has changed, the
`ProxyMixin.set_state_of_channels()` method will still raise an
exception, as expected.