feat(proxy/channels): add `append` kwarg
Prior to this commit, if a set of channel states was specified as a
`pandas.Series` instance to the `ProxyMixin.set_state_of_channels()`
method, the corresponding channel states were _merged_ with the existing
actuation states.  In order to _overwrite_ any existing actuation
states, the corresponding channel states needed to be reset explicitly,
e.g., with a call to the `turn_off_all_channels()` method.

To reduce boilerplate code along with the number of required serial
communication calls, add the optional `append` keyword argument to
the `ProxyMixin.set_state_of_channels()` method.  If `True`, append
states specified as a `pandas.Series` instance. Otherwise, overwrite
the existing channel states.