feat(move): add move_results_to_frame()
The results from the `dropbot.move.move_liquid()` function may be easily
serialized as JSON. However, when attempting to, for example, plot the
results, it is easier to manage the data in a data frame.

In this commit, add the `move_results_to_frame()` function to the
`dropbot.move` module to convert results from `move_liquid()` to a data
frame.