fix(init): explicitly import core.__version__
Prior to this commit, attributes of the `core` module were exposed
through `__init__` using the following import statement:

```python
from .core import *
```

However, by default, `import *` imports all names _except_ those
beginning with an underscore (`_`).  This means that the `__version__`
attribute (which was originally accessible as an attribute of
`__init__`) was no longer exposed.

In this commit, explicitly import `core.__version__` in `__init__`.