fix(node): fix handling of <120 channels
Prior to this commit, getting and setting of the state of channels only
worked as expected when 120 channels were detected (i.e., 3 switching
boards).  However, the DropBot _should_ be able to operate in
configurations with fewer than 3 switching boards.  The getter for the
disabled channels mask was also affected similarly.

In this commit, fix the following methods to support fewer than 120
channels:

 - `Node::state_of_channels()`
 - `Node::set_state_of_channels()`
 - `Node::disabled_channels_mask()`