fix(voltage-source): clamp `pot/wiper_value`
Prior to this commit, since dc202c971b696e598b2e05eec29bf45bb3b296c6,
`wiper_value` was cast as a `uint8_t` instead of `float`, causing values
over 255 to overflow. This resulted in a high voltage being applied when
a voltage below the minimum was set **(VERY DANGEROUS)**.

In this commit, clamp out-of-range values for `pot_value` and
`wiper_value` (fixes #44).